Fix dynamic framework/modules build error
When using TLLayoutTransitioning as a CocoaPod where `use_frameworks!`
is specified and Modules are enabled, the framework fails to build
because of how the easing.h header from AHEasing is being imported:

“Include of non-modular header inside framework module
‘TLLayoutTransitioning.UICollectionView_TLTransitioning’”

This is easily resolved by changing the AHEasing import to a framework
style import.

See
http://blog.cocoapods.org/Pod-Authors-Guide-to-CocoaPods-Frameworks/
for further discussion.