Fix dynamic framework/modules build error
The previous commit wasn’t actually sufficient, what we need to do is
use the new `@import` syntax to import AHEasing as a module if modules
are enabled. If modules aren’t enabled, we’ll just fall back to the
original textual import of the `easing.h` file.

In order to prevent a build error within the TLLayoutTransitioning
project, I have also disabled modules in the project. (Probably would
be a good idea to create a Podfile for the project itself so that
AHEasing can be pulled in as a dependency using CocoaPods, which would
let this project build.)